+++ backend/schema.sql
@@ -1,5 +1,7 @@
 create table diary (date date not null, id integer generated by default as identity, user_id bigint not null, body TEXT not null, title varchar(255) not null, primary key (id));
+create table taak_task (id integer generated by default as identity, is_all_day boolean not null, load_score integer not null, completed_at timestamp(6), due_at timestamp(6) not null, user_id bigint not null, memo TEXT not null, title varchar(255) not null, primary key (id));
 create table taak_user (id bigint generated by default as identity, line_sub varchar(255) unique, nick_name varchar(255), password varchar(255), username varchar(255) unique, primary key (id));
 create table taak_user_roles (taak_user_id bigint not null, roles varchar(255));
 alter table if exists diary add constraint FKicd4egl46ha9186c3qi4f7kh7 foreign key (user_id) references taak_user;
+alter table if exists taak_task add constraint FKnx44vkinsf3rvymxg6em30dp9 foreign key (user_id) references taak_user;
 alter table if exists taak_user_roles add constraint FKoahoj7by44rbh3ke1txr22udo foreign key (taak_user_id) references taak_user;
