+++ backend/src/main/java/com/team8/taaks/model/TaakTaskRepository.java
@@ -1,4 +1,4 @@
-package com.team8.taak.model;
+package com.team8.taaks.model;
 
 import java.util.List;
 import java.util.Optional;
