+++ backend/src/test/java/com/team8/taaks/JdbcConnectionTest.java
@@ -1,4 +1,4 @@
-package com.team8.taak;
+package com.team8.taaks;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
